all mappings and repos ready

+++ server/src/main/java/com/cmpe275/cartpool/repos/ProductStoreRepo.java
@@ -3,5 +3,5 @@ package com.cmpe275.cartpool.repos;
 import com.cmpe275.cartpool.entities.ProductStore;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductStoreRepo extends JpaRepository<ProductStore,Long> {
+public interface ProductStoreRepo extends JpaRepository<ProductStore,Integer> {
 }
